requirements.txt
numpy == 1.18.4
opencv-python == 4.2.0.34
opencv-contrib-python == 4.2.0.34
pyqt5 ==5.15.0
openpyxl == 3.0.4
gspread == 3.6.0
pillow == 7.1.2
imutils == 0.5.3
sklearn == 0.0
matplotlib == 3.3.3
tensorflow == 2.3.1
keras == 2.3.1
dlib == 19.21.0
serial == 0.0.97